Reduce Brewfile
Dependencies (like antigen) must be installed before restoring dotfiles.
Having all formulae and casks I've ever used in the Brewfile complicates
this process. So, I'm reducing it to the core things I'll always want
installed.

The next step would be to update the init script to run `brew bundle`
and verifying that all dependencies are in place.